requirements.txt
openpyxl
langchain
numpy
pandas
openai
langchain_openai
requests
langchain-community
pypdf
python-dotenv
faiss-cpu
psycopg2-binary
pydantic
ftfy
asyncpg
fastapi
fastapi-sessions
motor
uvicorn[standard]
sqlalchemy
